chore: reorganise code out of main
